+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 - You can now specify a `hybrid_index` when creating an index for the document index to use hybrid (semantic and keyword) search.
 - `min_score` and `max_results` are now optional parameters in `DocumentIndexClient.SearchQuery`.
 - `k` is now an optional parameter in `DocumentIndexRetriever`.
+- List all indexes of a namespace with `DocumentIndexClient.list_indexes`.
+- Remove an index from a namespace with `DocumentIndexClient.delete_index`.
 
 ### Fixes
-...
+- `DocumentIndexClient` now properly sets `chunk_overlap` when creating an index configuration.
+
 ### Deprecations 
 ...
 ### Breaking Changes